refactor: split base.py into drivers.py and parser.py
