+++ src/xplane_apt_convert/__init__.py
@@ -1,0 +1,3 @@
+from .parser import ParsedAirport
+
+__version__ = "0.5.7"
