--- xplane_apt_convert/__init__.py
@@ -1,3 +1,0 @@
-from .base import ParsedAirport
-
-__version__ = "0.5.7"
